Move internal class to package de.fayard.internal
