+++ plugin/src/main/kotlin/de/fayard/internal/EditorConfig.kt
@@ -1,4 +1,4 @@
-package de.fayard
+package de.fayard.internal
 
 import java.io.File
