+++ plugin/src/main/kotlin/de/fayard/internal/OutputFile.kt
@@ -1,5 +1,6 @@
-package de.fayard
+package de.fayard.internal
 
+import de.fayard.BuildSrcVersionsExtension
 import org.gradle.api.Project
 
 internal enum class OutputFile(var path: String, var existed: Boolean = false, val alternativePath: String? = null) {
@@ -17,7 +18,7 @@ internal enum class OutputFile(var path: String, var existed: Boolean = false, v
     }
 
     fun logFileWasModified() {
-        Companion.logFileWasModified(path, existed)
+        logFileWasModified(path, existed)
     }
 
     companion object {
