+++ sample-kotlin/gradle.properties
@@ -1,15 +1,15 @@
 # Generated by $ ./gradlew refreshVersions
 # See https://github.com/jmfayard/buildSrcVersions
 version.org.jetbrains.kotlin.jvm.gradle.plugin=1.3.11
-                               # \-- available=1.3.50
+                                   # available=1.3.50
 version.kotlin.scripting.compiler.embeddable=1.3.11
-                             # \-- available=1.3.50
+                                 # available=1.3.50
 version.com.gradle.build.scan.gradle.plugin=2.4.1
-                            # \-- available=2.4.2
+                                # available=2.4.2
 version.guava=15.0
-# \-- available=23.0
+  # available=23.0
 version.guice=2.0
-# \-- available=4.2.2
+  # available=4.2.2
 # Plugin versions
 # See https://github.com/jmfayard/buildSrcVersions/issues/60
 plugin.org.jetbrains.kotlin.jvm=1.3.11
