+++ sample-versionsOnlyMode/gradle.properties
@@ -1,3 +1,16 @@
+# Generated by $ ./gradlew refreshVersions
+# See https://github.com/jmfayard/buildSrcVersions
+version.org.lovedev.greeting.kotlin.gradle.plugin=1.1
+version.org.jetbrains.kotlin.jvm.gradle.plugin=1.3.50
+version.ch.tutteli.kotlin.utils.gradle.plugin=0.29.0
+version.com.gradle.build.scan.gradle.plugin=2.4.1
+                                # available=2.4.2
+version.nebula.kotlin.gradle.plugin=1.3.50
+version.org.jetbrains.kotlin=1.3.50
+version.guava=15.0
+  # available=23.0
+version.guice=2.0
+  # available=4.2.2
 # Plugin versions
 # See https://github.com/jmfayard/buildSrcVersions/issues/60
 plugin.org.lovedev.greeting.kotlin=1.1
@@ -6,10 +19,3 @@ plugin.ch.tutteli.kotlin.utils=0.29.0
 plugin.com.gradle.build-scan=2.4.1
 plugin.nebula.kotlin=1.3.50
 # You can edit the rest of the file, it will be kept intact
-# Generated by $ ./gradlew refreshVersions
-# See https://github.com/jmfayard/buildSrcVersions
-version.org.jetbrains.kotlin.jvm.gradle.plugin=1.3.50
-version.org.jetbrains.kotlin=1.3.50
-version.okhttp=4.1.0
-# \-- available=4.1.1
-version.okio=2.0.0
